feat(category): Category/Hashtag 커맨드 리포지토리 및 AdminCategoryCommandService 구현
- CategoryCommandRepository, HashtagCommandRepository 포트와 JPA 기반
  구현체(CategoryCommandRepositoryImpl, HashtagCommandRepositoryImpl) 추가
- CategoryJpaRepository를 CategoryCommandJpaRepository로 대체
- AdminCategoryCommandService에 카테고리 생성, 해시태그 수동 연결/해제,
  병합요청 승인/반려, 해시태그 삭제 커맨드 구현
- 해시태그 연결·승인·반려 상태 관리 책임을 Category 애그리거트에서
  CategoryHashtag 엔티티로 위임
- CategoryHashtag.similarityScore 타입을 Double에서 BigDecimal로 변경

+++ apps/product-service/src/main/java/com/sub9/productservice/category/domain/entity/CategoryHashtag.java
@@ -18,6 +18,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -50,7 +51,7 @@ public class CategoryHashtag extends BaseEntity {
     private CategoryHashtagStatus status;
 
     @Column(name = "similarity_score", precision = 5, scale = 4, nullable = false)
-    private Double similarityScore;
+    private BigDecimal similarityScore;
 
     @Column(name = "unique_version", nullable = false)
     private UUID uniqueVersion;
@@ -67,7 +68,7 @@ public class CategoryHashtag extends BaseEntity {
             Hashtag hashtag,
             CategoryHashtagMatchType matchType,
             CategoryHashtagStatus status,
-            Double similarityScore
+            BigDecimal similarityScore
     ) {
         return CategoryHashtag.builder()
                 .category(category)
@@ -99,7 +100,7 @@ public class CategoryHashtag extends BaseEntity {
             Hashtag hashtag,
             CategoryHashtagMatchType matchType,
             CategoryHashtagStatus status,
-            Double similarityScore,
+            BigDecimal similarityScore,
             UUID uniqueVersion
     ) {
         this.category = category;
